Fix generate root certificate at the first time
The frontend want to receive an object as the list of custom names of
root certificate and when is before the first generation of root certificate,
the list with custom names is empty and PHP return an array. With this
change every time will send an object instead of array when is an empty
list.

Signed-off-by: Vitor Mattos <vitor@php.rio>