Make possible change renderAs
To be possible customize the renderAs argument of TemplateResponse
class, I implemented the same of was made to customize the template
name. In the future, maybe will be good replace this by specific error class,
an error class to each kind of template. But for now maybe this was the
most simple way that I found to solve the problem.

Signed-off-by: Vitor Mattos <vitor@php.rio>