Make possible save a file content into a file of different authenticated user
Problem: Nextcloud server disalowed to write a content into a file that isn't of authenticated user.

Workaround: to prevent error when try to save a file in a folder of different authenticated user

At the follow code:
https://github.com/nextcloud/server/blob/4173dfe05bd0155eb217dd428ac82091a508567a/apps/files_versions/lib/Listener/FileEventsListener.php#L350-L366
Nextcloud server force to use the user folder to get the file of authenticated user.
This piece of code is to bypass the logic to use the authenticated user.

@todo Identify a way to be possible save the file content

Signed-off-by: Vitor Mattos <vitor@php.rio>