Prevent changes at initial dataProvider
The vfsStream get crazy behavior when we run all tests. If we execute
a test isolated, will work fine, if run together with all, the file
don't will exstists. I identified that the dataProvider is started
before all and maybe other tests that use vfs is affecting this tests
and to prevent the problem, I created the vfs to each test.

Signed-off-by: Vitor Mattos <vitor@php.rio>