chore: reset modal when load and close
The certificate data, password, error message need to be empty when
modal is open. I also put this to close function to prevent to have
these data when is no more necessary.

I also converted the modal to large when display the certificate data
because in some cases, the certificate data have very large rows and a
lot of rows.

Signed-off-by: Vitor Mattos <vitor@php.rio>