feat: allow to change signature hash algorithm
The default hash algorithm of JSignPdf is SHA1 and with this change was
defined SHA256 as default hash algorithm and make possible choose
between the allowed algorithms: SHA1, SHA256, SHA384, SHA512, RIPEMD160.

To don't throw the follow message, was implemented a logic to identify the
right algorithm to sign the document depending of PDF version but the
implemented algorithm usign the prefix of PDF file, isn't effective. The
best way to identify the pdf version is using the command pdfinfo of
poppler-utils.

Signed-off-by: Vitor Mattos <vitor@php.rio>