fix: migration column
When was published the biggest release 8 of LibreSign, was made a
mistake and the column name was changed causing inconsistent data at old
registers.

Because haven't backup of old data, I made a workaround to restore the
possible data considering that at migration Version8000Date20230420125331.php
the field `signed` was used to fill identified_at_date and
last_attempt_date.

This fix only will affect the problematic rows.

Signed-off-by: Vitor Mattos <vitor@php.rio>