fix: user userId when validate file
Every when we load a file from Nextcloud, would be good to have the
fileId to be quicky. When we request to sign a file from fileId, the
file already exists before create the sign request and because this, we
haven't a LibreSign file. To prevent error, I get the userId of who
requessted to sign the document and sent to validation methods and after
whis, was possible to validate the file.

Signed-off-by: Vitor Mattos <vitor@php.rio>