fix: fallback when system haven't a TTF font
Before this, when we open `Administration Settings > LibreSign` and the installation of extension Imagick haven't access to any ttf font, throws the follow error:

```xml
<?xml version="1.0"?>
<ocs>
 <meta>
  <status>failure</status>
  <statuscode>400</statuscode>
  <message></message>
 </meta>
 <data>
  <error>No fonts available at system</error>
 </data>
</ocs>
```

As fallback, I added the packed DejaVuSerif font and implemented an unit test to verify if mpdf have this font.

I used this font because we already pack this font with LibreSign and the test is to prevent problems if this font is renamed, moved or removed from mpdf.

Signed-off-by: Vitor Mattos <vitor@php.rio>