fix: prevent error when check binaries of JSignPdf
Every when the message "Invalid hash of binaries files" is throws, is added
a log with the reason. At this case was consequence of previous change
at setup of JSignPdf when was added a folder called jsignpdf_home to
prevent warning of JSignPdf when is executed at CLI mode.

The folder in question is `conf`, and with this, will be created into
another place.

At the log I got this:

```json
{
    "EXTRA_FILE": {
        "conf/conf.properties": {
            "expected": "",
            "current": "cf83e1357eefb8bdf1542850d66d8007d620e4050b5715dc83f4a921d36ce9ce47d0d13c5d85f2b0ff8318d2877eec2f63b931bd47417a81a538327af927da3e"
        },
        ".JSignPdf": {
            "expected": "",
            "current": "cf83e1357eefb8bdf1542850d66d8007d620e4050b5715dc83f4a921d36ce9ce47d0d13c5d85f2b0ff8318d2877eec2f63b931bd47417a81a538327af927da3e"
        }
    }
}
```

This don't affect LibreSign but could generate a false image that could
not work fine.

Signed-off-by: Vitor Mattos <vitor@php.rio>