fix: store the date that the file was signed
The sign engine is used to sign the file and between the interval that
the file was signed and the moment to save this at the database, is
possible to be different. To prevent issues about different dates
(seconds or milisseconds), I made a parse at signed file and after then,
I fetch the last signature date to store this at signed field at the
database.

Signed-off-by: Vitor Mattos <1079143+vitormattos@users.noreply.github.com>