chore: cover the sign method using pkcs7 engine
When we sign a PDF we can use PKCS12, but when we sign other files, is
necessary to use PKCS7.

I added xml as file extension to be possible check if the signature is
made with success using a file different of PDF.

Note: For now we haven't an entry point to sign a non PDF file, this is
only a concept and is necessary to implement the changes at frontend and
at API side too to be possible sign usign non PDF files.

Signed-off-by: Vitor Mattos <1079143+vitormattos@users.noreply.github.com>