Fix ConfigurationCheck failure on poppler-utils
By default, PHP-FPM clears all environment variables in workers. The result of this is that the "which" command, previously used to determine that pdfsig and pdfinfo commands are available on the system would return NULL instead of the path to the commands.

We solve this by instead simply executing the command "pdfinfo -v" or "pdfsig -v" and checking the return value. A return value of anything besides "0" indicates that the command failed.

Signed-off-by: Adam Serbinski <37986043+ASerbinski@users.noreply.github.com>