fix: remove condition that restricts LibreSign tab to LibreSign files
In the Files app, users should be able to open the LibreSign tab even
with non-LibreSign files to start the sign flow.

The condition was introduced in the last release because the same
sidebar tab is used in the LibreSign app, where the file is always a
LibreSign file.

To reproduce this:

Go to app Files
Upload a new PDF file
Click over this file and go to open in LibreSign file

The expected is to see the button "Add signer"

Signed-off-by: Vitor Mattos <1079143+vitormattos@users.noreply.github.com>