fix: prevent error when try to create a folder two times
The LibreSign folder is created by every event that need to fetch files
from this folder. When we try to create a folder or a file, before to
finish the creation proccess, the folder is locked by Nextcloud. To
prevent this, I made first a get, if is not found, I try to create.

Signed-off-by: Vitor Mattos <1079143+vitormattos@users.noreply.github.com>