fix: prevent error when send reminders
LibreSign stores the username of each account responsible for signing.
When checking for pending signatures to send reminders, the query could
return a user that no longer exists in Active Directory. Although LibreSign
is designed to ignore deleted accounts, it cannot detect when an account
simply disappears (i.e., is no longer returned by AD). As a result, the
process stopped when it encountered the missing user.

Added an error handler to skip such cases and continue processing the
remaining notifications. This fix will be included in the next release.

Signed-off-by: Vitor Mattos <1079143+vitormattos@users.noreply.github.com>