fix: update modal title to reflect edit vs add signer action
The modal dialog was displaying 'Add new signer' title when clicking
an existing signer to edit, which was confusing. Added a computed
property 'modalTitle' that dynamically returns 'Edit signer' when
editing an existing signer or 'Add new signer' when adding a new one.

The logic checks if signerToEdit object has properties to determine
the correct context.

Signed-off-by: Vitor Mattos <1079143+vitormattos@users.noreply.github.com>